predict damage examine
