+++ Content.Client/Damage/Systems/DamageOtherOnHitSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared.Damage.Systems;
+
+namespace Content.Client.Damage.Systems;
+
+public sealed class DamageOtherOnHitSystem : SharedDamageOtherOnHitSystem;
