+++ Content.Shared/Damage/Components/DamageOtherOnHitComponent.cs
@@ -1,0 +1,24 @@
+using Content.Shared.Damage.Systems;
+
+namespace Content.Shared.Damage.Components;
+
+/// <summary>
+/// Makes this entity deal damage when thrown at something.
+/// </summary>
+[RegisterComponent]
+[Access(typeof(SharedDamageOtherOnHitSystem))]
+public sealed partial class DamageOtherOnHitComponent : Component
+{
+    /// <summary>
+    /// Whether or not to ignore damage modifiers.
+    /// </summary>
+    [DataField]
+    public bool IgnoreResistances = false;
+
+    /// <summary>
+    /// The damage amount to deal on hit.
+    /// </summary>
+    [DataField]
+    public DamageSpecifier Damage = default!;
+
+}
